Move drive change pdf gen into separate file
